requirements.txt
fastapi~=0.115.2
uvicorn~=0.32.0
sqlalchemy~=2.0.36
databases~=0.9.0
python-dotenv~=1.0.1
psycopg2-binary~=2.9.10
typing-extensions~=4.12.2
supabase~=2.9.1
PyJWT~=2.9.0